Make Ghidra decompileFunction timeout configurable
This commit adds "Decompile Timeout" option to the MCP plugin that
allows for setting a custom timeout to the `decompileFunction`. In
particular it is useful to set to 0 or a large value when dealing with
large functions that take a while to decompile.

The default value is set to 30.

In addition as part of this commit, the request timeout from the MCP
bridge to the plugin is made configurable with the `--ghidra-timeout`
argument. The default value is 5 seconds.